+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelogs
 
+## 2.2.27
+
+- added fake profiles for SG/PR with no local addresses
+- added fake profile for foreigner with blank address fields
+- added fake profile for foreigner with absent address object
+
 ## 2.1.0
 
 - Using mol-lib-config